milvus/requirements.txt
boto3==1.22.11
botocore==1.25.12
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
googleapis-common-protos==1.56.0
grpcio==1.46.0
grpcio-status==1.46.0
grpcio-tools==1.37.1
idna==3.3
jmespath==1.0.0
mmh3==3.0.0
numpy==1.22.3
opencv-python==4.5.5.64
pandas==1.4.2
Pillow==9.1.0
pkg_resources==0.0.0
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymilvus==2.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
rsa==4.8
ruamel.yaml==0.16.6
s3transfer==0.5.2
six==1.16.0
tabulate==0.8.9
timm==0.5.4
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
towhee==0.6.1
tqdm==4.64.0
typing_extensions==4.2.0
ujson==5.1.0
urllib3==1.26.9
